* Move metadata API out of dubbo-config-api
