--- dubbo-config/dubbo-config-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.deploy.ApplicationDeployListener
@@ -1,1 +1,0 @@
-exporter=org.apache.dubbo.config.metadata.ExporterDeployListener
