--- dubbo-config/dubbo-config-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.registry.client.ServiceInstanceCustomizer
@@ -1,1 +1,0 @@
-metadata-url=org.apache.dubbo.config.metadata.MetadataServiceURLParamsMetadataCustomizer
