+++ dubbo-registry/dubbo-registry-api/src/main/java/org/apache/dubbo/registry/deploy/ConfigurableMetadataServiceExporter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.config.metadata;
+package org.apache.dubbo.registry.deploy;
 
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.logger.ErrorTypeAwareLogger;
